Create a GDB pretty printer for Eterms
Reimplements much of the etp_commands script file and
etp.py to create a pretty printer for GDB. This is useful
to be able to see contents of Eterms in the variables panel in an IDE
(e.g. vscode). Also just using print (p) arg  instead of etp arg.

The GDB MI contains a bug which prevents us from displaying the
variable data in vscode, resolved by patching gdb,
see erts/etc/unix/etp_gdb_pp/README.md for more.
A launch.json configuration for vs code is available as well.

Supporting both the print statement and the MI interface
is problematic when it comes to formatting so two
pretty printers have been created.