stdlib: completion improvements
Add support for binary and string keys matches inside maps and functions
Display return type of functions in the shell?
Display registered names
If you've started an expression with either an atom,list,tuple,map,
string or binary after a ',' or a '->'
then we can assume that you want to do a pattern match.
If we are inside a function head then we look at the functions type spec
if we are inside a case, then we look at the case expression
if we are inside a receive, then we don't know.
if we are inside a function, fun, or simple shell expression, then we can
try to guess the user wants to pattern match against something in the shell
bindings.
If we already have an = sign to the right of the expression with a binding
 or a function call, then we can autocomplete based on the binding or function type.

Other ideas that probably needs a separate PR, but I'll mention them here:
Fish style completion after the cursor
Hooks for completion of a function parameter, based on the type or user customization
 e.g. output module names of function names (of a module) when the type is module()
      or function() (and previous argument is module), or when you want a process label
A user api for creating a custom prompt mode, like search, help, gpt etc.
A user api for creating a custom function on a key bindings